backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.4.2
sentence-transformers==2.7.0
groq==0.8.0
pypdf2==3.0.1
python-docx==1.1.2
numpy==1.26.4
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
